fix(mesas): dinero en BigDecimal, frontera offline propia y outbox write-ahead
Tres defectos P1 que encontro una auditoria independiente:

- 17 campos de dinero estaban en Double. 0.10 + 0.20 no da 0.30, y esto termina
  en el cheque de un cliente. Ahora BigDecimal, con el limite de Gson verificado
  (si el JSON pasa por Double antes, la precision ya se perdio y el fix no sirve).
- orQueueOffline reusaba el clasificador de PAGOS, que manda 401/403/500 y
  cualquier error DESCONOCIDO a reintentar. En Mesas eso significa que un
  NullPointerException nuestro se disfraza de 'exito offline' y el mesero ve una
  ronda guardada que no existe. Ahora Mesas tiene su propio clasificador: encola
  solo sin-red y 502/503/504/408/429; propaga todo lo demas.
- El outbox llamaba al server ANTES de crear el intent. La llave idempotente
  ahora existe antes del primer intento y se usa online y offline.

El clasificador de pagos (SyncOutcome.kt) queda intacto a proposito: su
semantica permisiva es correcta para dinero ya cobrado.

+++ app/src/main/java/com/jaac/avoqado_tpv/features/tables/data/api/dto/OpenTableDto.kt
@@ -1,5 +1,7 @@
 package com.jaac.avoqado_tpv.features.tables.data.api.dto
 
+import java.math.BigDecimal
+
 /**
  * `POST /tpv/venues/{venueId}/tables/{tableId}/open` (`tableController.openTable`,
  * re-export de `table.mobile.controller.ts::openTable`). Body: `{ covers?: number }`
@@ -28,5 +30,5 @@ data class OpenedOrder(
     val id: String,
     val orderNumber: String? = null,
     val version: Int = 1,
-    val total: Double? = null,
+    val total: BigDecimal? = null,
 )
